chore(vision): revert accidental file rename
